# server/docs/swagger.yaml
basePath: /v1
definitions:
...
  controller.EmailConfirmedRequest:
    properties:
      code:
        example: 54JK64
        type: string
      email:
        example: jon.doe@gmail.com
        type: string
    required:
    - code
    - email
    type: object
  controller.SendConfirmedEmailCodeRequest:
    properties:
      email:
        example: jon.doe@gmail.com
        type: string
    required:
    - email
    type: object
...
  controller.UserSignInRequest:
    properties:
      email:
        example: jon.doe@gmail.com
        type: string
      login:
        example: user1
        maxLength: 50
        minLength: 1
        type: string
      password:
        example: SuperPassword123
        type: string
    required:
    - password
    type: object
...
  response.Response:
    description: Structure for a standard API response
    properties:
      data: {}
      error:
        example: any error
        type: string
      status:
        example: success/error
        type: string
    type: object
host: film-catalog-8re5.onrender.com
info:
  contact:
    name: Evdokimov Igor
    url: https://t.me/epelptic
  description: API for potatorate site
  title: Film-catalog API
  version: 1.0.0
paths:
...
  /auth/{provider}:
    get:
      consumes:
      - application/json
      description: Redirects the user to the OAuth provider for authentication.
      parameters:
      - description: OAuth provider DON'T WORK IN SWAGGER!!!
        example: '"google" or "yandex"'
        in: path
        name: provider
        required: true
        type: string
      produces:
      - application/json
      responses:
        "307":
          description: Redirecting to provider
        "400":
          description: Provider not supported
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: User SignWithOauth
      tags:
      - auth
  /auth/{provider}/callback:
    get:
      consumes:
      - application/json
      description: Handles the callback from the OAuth provider after the user has
        authorized the app. THIS ENDPOINT IS CALLED BY THE OAUTH PROVIDER, NOT THE
        FRONTEND!!!
      parameters:
      - description: OAuth provider
        example: '"google" or "yandex"'
        in: path
        name: provider
        required: true
        type: string
      - description: State parameter sent during OAuth authorization
        example: '"randomstate123"'
        in: query
        name: state
        required: true
        type: string
      - description: Authorization code returned by OAuth provider
        example: '"authorizationcode123"'
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User already exists, successfully authenticated
          schema:
            $ref: '#/definitions/response.Response'
        "201":
          description: New user created and successfully authenticated
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Provider not supported or invalid state
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: OAuth2 Callback Handler
      tags:
      - auth
...
  /auth/refresh-token:
    post:
      consumes:
      - application/json
      description: Refreshes the access token using the provided refresh token from
        cookies.
      produces:
      - application/json
      responses:
        "200":
          description: Successfully refreshed access token
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Invalid, missing or expired refresh token
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Refresh Access Token
      tags:
      - auth
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: Create access and refresh token and return them to the user
      parameters:
      - description: User login details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/controller.UserSignInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User successfully signed in
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid request payload or validation error
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Invalid Password or Email
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: User email is not confirmed
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: User SignIn
      tags:
      - auth
...
  /email/confirm:
    put:
      consumes:
      - application/json
      description: Validate confirmed code and is it confirmed update email_status
      parameters:
      - description: data for confirmed email
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.EmailConfirmedRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success email confirmation
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Error email confirmation
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Confirmation email address
      tags:
      - email
  /email/send-code:
    post:
      consumes:
      - application/json
      description: Generate code for confirmation email and send this to email. This
        endpoint have rate 1 req in 1 min
      parameters:
      - description: Email пользователя для подтверждения
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.SendConfirmedEmailCodeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Код подтверждения успешно отправлен
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Ошибка валидации или неверный запрос
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.Response'
      summary: Send code for confirmation email
      tags:
      - email
...
schemes:
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"